llm-eval/requirements.txt
langchain
langchain-ollama
langchain-core
langsmith
ragas
python-dotenv
datasets
pypdf
ollama